requirements.txt
aiofiles==23.2.1
annotated-types==0.7.0
anyio==3.7.1
chardet==5.2.0
click==8.1.8
et_xmlfile==2.0.0
fastapi==0.104.1
h11==0.14.0
idna==3.10
Jinja2==3.1.2
lxml==5.3.1
MarkupSafe==3.0.2
numpy==2.2.4
openpyxl==3.1.2
pandas==2.2.3
pillow==11.1.0
pydantic==2.4.2
pydantic_core==2.10.1
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-docx==1.0.1
python-dotenv==1.0.0
python-multipart==0.0.6
python-pptx==0.6.21
pytz==2025.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.23
starlette==0.27.0
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.2
uvicorn==0.24.0
XlsxWriter==3.2.2 